+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Added trade notifications, which can be customized in the config.
 - Added global ban option in the config. This will ban pokemons like plushies globally.
 - Added history command to view past actions taken by the mod.ç
-- History supports databases: MongoDB.
+- History supports databases: MongoDB. (JSON implemented test this before using it in production.)
 
 ### Bug Fixes